Text_Steganography/requirements.txt
darkdetect==0.8.0
packaging==24.0
pillow==10.3.0
